+++ hyp3_gamma/util.py
@@ -10,6 +10,28 @@ log = logging.getLogger(__name__)
 gdal.UseExceptions()
 
 
+class GDALConfigManager:
+    """Context manager for setting GDAL config options temporarily"""
+    def __init__(self, **options):
+        """
+        Args:
+            **options: GDAL Config `option=value` keyword arguments.
+        """
+        self.options = options.copy()
+        self._previous_options = {}
+
+    def __enter__(self):
+        for key in self.options:
+            self._previous_options[key] = gdal.GetConfigOption(key)
+
+        for key, value in self.options.items():
+            gdal.SetConfigOption(key, value)
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        for key, value in self._previous_options.items():
+            gdal.SetConfigOption(key, value)
+
+
 def get_granule(granule):
     download_url = get_download_url(granule)
     zip_file = download_file(download_url, chunk_size=10485760)
@@ -33,11 +55,12 @@ def earlier_granule_first(g1, g2):
     return g2, g1
 
 
-def set_pixel_as_point(tif_file):
+def set_pixel_as_point(tif_file, shift_origin=False):
     ds = gdal.Open(tif_file, gdal.GA_Update)
-    transform = list(ds.GetGeoTransform())
-    transform[0] += transform[1] / 2
-    transform[3] += transform[5] / 2
-    ds.SetGeoTransform(transform)
     ds.SetMetadataItem('AREA_OR_POINT', 'Point')
-    ds = None
+    if shift_origin:
+        transform = list(ds.GetGeoTransform())
+        transform[0] += transform[1] / 2
+        transform[3] += transform[5] / 2
+        ds.SetGeoTransform(transform)
+    del ds