+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.0](https://github.com/ASFHyP3/hyp3-gamma/compare/v4.1.2...v4.2.0)
+
+### Added
+* Option for RTC_GAMMA jobs to use either Copernicus GLO-30 Public DEM or legacy SRTM/NED DEMs for processing:
+  * `--dem-name` option to `rtc` entrypoint and `rtc_sentinel.py` script
+  * `dem_name` parameter to `hyp3_gamma.rtc_sentinel.rtc_sentinel_gamma()`
+* `hyp3_gamma.dem` module for preparing GeoTIFF mosaics of the
+  [Copernicus GLO-30 Public DEM](https://registry.opendata.aws/copernicus-dem/)
+
+### Changed
+* Upgraded to hyp3-metadata [v0.4.2](https://github.com/ASFHyP3/hyp3-metadata-templates/blob/develop/CHANGELOG.md#042)
+  from v0.2.0
+
 ##  [4.1.2](https://github.com/ASFHyP3/hyp3-gamma/compare/v4.1.1...v4.1.2)
 
 ### Changed